move cleanNote to ForumLibrary
